requirements.txt
flask
python-dotenv
psycopg2
Flask-SQLAlchemy